requirements.txt
# DashScope (Alibaba Bailian) edition
langgraph==0.2.39
langchain==0.3.7
langchain-community==0.3.7
chromadb==0.5.11
pydantic==2.9.2
python-dotenv==1.0.1
pypdf==4.3.1
faiss-cpu==1.8.0.post1
typer==0.12.5
rich==13.9.2
requests==2.32.3
dashscope==1.24.6
